requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
django-environ==0.10.0
djangorestframework==3.14.0
psycopg==3.1.8
psycopg2==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.5.0
tzdata==2023.3
